Added functionality to manage images and captions
This commit introduces a series of methods in the main file to manage images and their respective captions. Also included are regex checks for file types and error messages when a related caption or image does not exist.

Changes were also made to the entities file where Image and Caption structs were modified to include extension and directory attributes. The Content and NumberOfTags attributes are currently commented out, but might be used in future iterations.