Refactor file reading using filepath.Walk
Replaced os.ReadDir with filepath.Walk for directory traversal in main.go. This improves performance as it enables concurrent reading of directories and files. Also enhances the handling of image and caption files by storing directory and extension information in Image and Caption structs.

The new approach also allows efficient checking of file types with regex and provides useful error messages if matching caption or image file cannot be found. Commentary also added for potential inclusion of Content and NumberOfTags attributes in the future.