Add method to replace spaces in captions with underscores
A new method, replaceSpaces, was implemented in methods.go. This was done to handle cases where spaces in captions need to be replaced with underscores. Addition of certain options in the menu in main.go was also performed to allow users to choose this new functionality. This change will help in standardizing the naming conventions and improving searchability of files.