Refactor log message printing and update dependencies
Updated the message printing logic in the main service to use a seemingly more readable and maintainable approach. Instead of sequentially calling the logger with each message, it now uses a printable array that contains all the messages and iterates over it.

In addition, updated 'roggy' and 'stemp' packages to versions v0.1.6 and v0.1.5 respectively. This ensures the application is using up-to-date packages, gaining any potential bugfixes or enhancements they may introduce. Also, potentially a necessary change to support the new message printing logic.