Use slice of functions for debug levels
This commit adds a map of logging functions to better handle different debug levels. Allowed levels are identified by numbers in strings, from '0' to '4'. Each number corresponds to a different function from the 'roggyPrinter' package and is used in the new 'printLog' function. In case an invalid level is provided, an error message is printed.