Add functionality to append new tags to captions
A new function 'appendNewTags' was added to 'caption-handlers.go' to append new tags to image captions. Furthermore, an option to call this function was added to the main.go switch case. The function is incomplete with a 'TODO' comment outlining the intended implementation and needs to be completed in the future.