Now reset the sender model after time expires, essentially making it as a temporary logger.
Also, `senderActiveDuration` is no longer a pointer as it's not necessary

Just remember to always think about updating the model and then returning the new model.

We're only spawning a new `tickMsg` if m.sender.Active is false, so only on the first run. We're free to reset the `senderActiveDuration` as necessary.