Added proper mutexes for locking map accesses to DataSet.Image and TempCaption.
singleDirMultiple() in `tui.go:30` now sends a `sender.ResultMsg` so that we log each directory that's being processed.
It has a WaitGroup for all directories to be processed with the sender.ResultMsg `tui.go:72:88`
Also, WaitGroup for `writeFilesMsg`, though this is technically unused right now. It gets constructed in `singleDirToMultiple` `tui.go:57`

Exported the `ImagesLock` and `CaptionLock` in the `DataSet` struct so that we can access these in any part of the code in `entities.go`

Each counter in `counters.go` now also lock the reads to the map

Also added a RWLock in methods.go:81,98 when accessing these slices.

Also sends a basic Refresh() msg in `update.go:291` when the logger ends to ensure updates whenever we do a dataset update

We're now showing the progress bar properly when adding directories to the dataset in `update.go:179`